DevSynch

+++ src/main/java/com/heliosapm/jvmti/agent/Agent.java
@@ -19,17 +19,22 @@ import java.io.File;
 import java.io.FileOutputStream;
 import java.io.InputStream;
 import java.lang.management.ManagementFactory;
+import java.lang.reflect.Field;
 import java.util.Arrays;
 import java.util.Collections;
+import java.util.Comparator;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import java.util.TreeSet;
 import java.util.concurrent.ThreadPoolExecutor;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicLong;
 
+import com.heliosapm.jvmti.util.SystemClock;
+import com.heliosapm.jvmti.util.SystemClock.ElapsedTime;
 import com.heliosapm.shorthand.attach.vm.VirtualMachine;
 
 
@@ -42,6 +47,10 @@ import com.heliosapm.shorthand.attach.vm.VirtualMachine;
  */
 
 public class Agent {
+	/** The singleton instance */
+	private static volatile Agent instance = null;
+	/** The singleton instance ctor lock */
+	private static Object lock = new Object();
 	/** Indicates if the native library is loaded */
 	private static final AtomicBoolean nativeLoaded = new AtomicBoolean(false);
 	/** The command line agent load prefixed */
@@ -51,6 +60,9 @@ public class Agent {
 	public static final String LIB_NAME = "oif_agent";
 	/** The system property override defining the absolute location of the oif agent library to load */
 	public static final String CONFIG_AGENT_LOCATION = "com.heliosapm.jvmti.lib";
+	/** The system property override to suppress the deleteOnExit on the jar extracted lib. Just set, no value needed. */
+	public static final String CONFIG_AGENT_NO_DEL_LIB = "com.heliosapm.jvmti.lib.nodelonexit";
+	
 	/** The directory prefix when loading the default lib in dev mode */
 	public static final String DEV_DIR_PREFIX = "target/native/";
 	/** This JVM's process id */
@@ -58,58 +70,88 @@ public class Agent {
 	/** The tag serial */
 	private final AtomicLong tagSerial = new AtomicLong(0L);
 	
+	/** Integer sorter */
+	@SuppressWarnings("unchecked")
+	public final Comparator<Integer> INT_COMP_ASC = (Comparator<Integer>) new TreeSet<Integer>().comparator();
+	/** Integer descending sorter */
+	public final Comparator<Integer> INT_COMP_DESC = Collections.reverseOrder(INT_COMP_ASC);
+	/**
+	 * Acquires the singleton agent instance
+	 * @return the agent
+	 */
+	public static Agent getInstance() {
+		if(instance==null) {
+			synchronized(lock) {
+				if(instance==null) {
+					instance = new Agent(true);
+				}
+			}
+		}
+		return instance;
+	}
+	
+	private Agent(final boolean loadNative) {
+		if(loadNative) {
+			loadNative();
+		}
+	}
 	
 	public static void main(String[] args) {
-		final Agent agent = new Agent();
 		log("Hello World");
-		agent.loadNative();
+		final Agent agent = new Agent(true);
 		log("Was Loaded:" + agent.wasLoaded());
 		int csCount = countInstances0(CharSequence.class, agent.tagSerial.incrementAndGet(), Integer.MAX_VALUE);
 		log("There are " + csCount + " CharSequence instances");
 		int a = agent.countExactInstances(Thread.class);
 		log("There are " + a + " instances of " + Thread.class);		
-       	Object[] objs = agent.getExactInstances(Thread.class);
+       	Object[] objs = agent.getInstancesOf(Thread.class);
        	log("Arr Length:" + objs.length);
        	log("Threads: " + java.util.Arrays.toString(objs));
-       	objs = agent.getExactInstances(ThreadPoolExecutor.class, 3);
+       	objs = agent.getInstancesOf(ThreadPoolExecutor.class, 3);
        	log("Arr Length:" + objs.length);
        	log("TPEs: " + java.util.Arrays.toString(objs));
-       	objs = agent.getExactInstances(System.out.getClass(), 300);
+       	objs = agent.getInstancesOf(System.out.getClass(), 300);
        	log("Arr Length:" + objs.length);
        	log("PrintStreams: " + java.util.Arrays.toString(objs));
-       	objs = agent.getExactInstances(String.class, 300);
+       	objs = agent.getInstancesOf(String.class, 300);
        	log("Arr Length:" + objs.length);
        	log("Strings: " + java.util.Arrays.toString(objs));
-       	objs = agent.getExactInstances(String[].class, 300);
+       	objs = agent.getInstancesOf(String[].class, 300);
        	log("Arr Length:" + objs.length);
        	log("String Arrays: " + java.util.Arrays.deepToString(objs));
        	log("int instance count: %s", agent.countExactInstances(byte[].class));
-       	agent.printClassCardinality(Object.class);
+       	
+       	log("==== Types of charsequence");
+       	for(Class<?> clazz: agent.getAllTypesOf(CharSequence.class)) {
+       		log("\t%s", clazz.getName());
+       	}
+       	
+       	final int loops = 20000;
+       	final ElapsedTime et = SystemClock.startClock();
+       	int max = 0;
+       	int min = Integer.MAX_VALUE;
+       	int maxLoop = -1;
+       	int minLoop = -1;
+       	for(int i = 0; i < loops; i++) {
+       		Object[] all = agent.getInstancesOfAny(Object.class);
+       		final int count = all.length;
+       		if(count > max) {
+       			max = count;
+       			maxLoop = i;
+       		}
+       		if(count < min) {
+       			min = count;
+       			minLoop = i;
+       		}
+       		all = null;
+       		if(i%1000==0) System.gc();
+       	}
+       	log(et.printAvg("AllObjects Lookup", loops));
+       	log("Max: %s at loop %s, Min: %s at loop %s", max, maxLoop, min, minLoop);
+       	
+       	//agent.printClassCardinality(Object.class);
 	}
 	
-	public void printClassCardinality(final Class<?> type) {
-		final Object[] objs = getInstances(type);
-		Map<Class<?>, Integer> card = classCardinality(objs);
-		log("======== Cardinality for type [%s] ========",type.getName());
-		long total = 0;
-		for(Map.Entry<Class<?>, Integer> entry: card.entrySet()) {
-			log("\t" + entry.getKey().getName() + ":" + entry.getValue());
-			total += entry.getValue();
-		}
-		log("======== Total: [%s]", total);
-	}
-	
-	public void printExactClassCardinality(final Class<?> type) {
-		final Object[] objs = getExactInstances(type);
-		Map<Class<?>, Integer> card = classCardinality(objs);
-		log("======== Cardinality for exact type [%s] ========",type.getName());
-		long total = 0;
-		for(Map.Entry<Class<?>, Integer> entry: card.entrySet()) {
-			log("\t" + entry.getKey().getName() + ":" + entry.getValue());
-			total += entry.getValue();
-		}
-		log("======== Total: [%s]", total);
-	}
 	
 	
 	public Map<Class<?>, Integer> classCardinality(final Object...instances) {
@@ -135,6 +177,19 @@ public class Agent {
 	}
 	
 	/**
+	 * Returns all loaded types equal to or inherrited from the passed type
+	 * @param type The type to retrieve
+	 * @return a set of classes
+	 */
+	public Set<Class<?>> getAllTypesOf(final Class<?> type) {
+		final Set<Class<?>> types = new HashSet<Class<?>>();
+		for(Object obj: getInstancesOfAny(type)) {
+			types.add(obj.getClass());
+		}
+		return types;
+	}
+	
+	/**
 	 * Determines the os and cpu arch from system properties and returns the 
 	 * directory and library file name
 	 * @return the directory and library file name for this platform
@@ -144,17 +199,17 @@ public class Agent {
 		final String arch = System.getProperty("sun.arch.data.model");
 		if(os.contains("linux")) {
 			if(arch.equals("64")) {
-				return "linux64/oif_agent.so";
+				return "linux64/liboifagent.so";
 			} else if(arch.equals("32")) {
-				return "linux32/oif_agent.so";
+				return "linux32/liboifagent.so";
 			} else {
 				throw new RuntimeException("Linux arch Not Supported: [" + arch + "]");
 			}
 		} else if(os.contains("windows")) {
 			if(arch.equals("64")) {
-				return "win64/oif_agent.dll";
+				return "win64/oifagent.dll";
 			} else if(arch.equals("32")) {
-				return "win32/oif_agent.dll";
+				return "win32/oifagent.dll";
 			} else {
 				throw new RuntimeException("Windows arch Not Supported: [" + arch + "]");
 			}			
@@ -195,7 +250,7 @@ public class Agent {
 			try {
 				// Check for sysprop override
 				final String libToLoad;
-				if(System.getProperties().contains(CONFIG_AGENT_LOCATION)) {
+				if(System.getProperties().containsKey(CONFIG_AGENT_LOCATION)) {
 					libToLoad = System.getProperties().getProperty(CONFIG_AGENT_LOCATION);									
 				} else {
 					// determine if we're running in a jar (true) or dev mode (false)
@@ -253,12 +308,16 @@ public class Agent {
 		int bytesRead = -1;
 		try {
 			final File libFile = File.createTempFile(LIB_NAME, libFileName);
-			libFile.deleteOnExit();
 			fos = new FileOutputStream(libFile);
 			while((bytesRead = is.read(transferBuffer))!=-1) {
 				fos.write(transferBuffer, 0, bytesRead);
 			}
 			fos.flush();
+			fos.close();
+			if(!System.getProperties().containsKey(CONFIG_AGENT_NO_DEL_LIB)) {
+				libFile.deleteOnExit();
+			}
+			
 			return libFile;
 		} finally {
 			if(is!=null) try { is.close(); } catch (Exception x) {/* No Op */}
@@ -318,7 +377,7 @@ public class Agent {
 	 * Negatives values will be rewarded with an {@link IllegalArgumentException}.
 	 * @return A [possibly zero length] array of objects
 	 */
-	public Object[] getExactInstances(Class<?> klass, int maxInstances) {
+	public Object[] getInstancesOf(final Class<?> klass, final int maxInstances) {
 		if(klass==null) throw new IllegalArgumentException("The passed class was null");
 		if(maxInstances<0) throw new IllegalArgumentException("Invalid maxInstances value [" + maxInstances + "]");
 		return getExactInstances0(klass, tagSerial.incrementAndGet(), maxInstances==0 ? Integer.MAX_VALUE : maxInstances);
@@ -329,8 +388,8 @@ public class Agent {
 	 * @param klass The class to search and return instances of
 	 * @return A [possibly zero length] array of objects
 	 */
-	public Object[] getExactInstances(Class<?> klass) {
-		return getExactInstances0(klass, tagSerial.incrementAndGet(), Integer.MAX_VALUE);
+	public Object[] getInstancesOf(final Class<?> klass) {
+		return getInstancesOf(klass, Integer.MAX_VALUE);
 	}
 	
 	/**
@@ -338,19 +397,72 @@ public class Agent {
 	 * @param klass The class to search and return instances of
 	 * @return A [possibly zero length] array of objects
 	 */
-	public Object[] getInstances(Class<?> klass) {
+	public Object[] getInstancesOfAny(final Class<?> klass) {
 		return getInstances0(klass, tagSerial.incrementAndGet(), Integer.MAX_VALUE);
 	}
 	
 	/**
-	 * Indicates if the agent was loaded or attached
+	 * Indicates if the agent was loaded or attached.
+	 * Useful as a basic test for the agent presence.
 	 * @return true if loaded, false if attached
 	 */
 	public boolean wasLoaded() {
-		return wasLoaded0();
+		try {
+			return wasLoaded0();
+		} catch (Throwable t) {
+			return false;
+		}
+	}
+
+	/**
+	 * Prints the count of each type equal to or inherrited 
+	 * from the passed type found in the heap
+	 * @param type The type to search the heap for
+	 * @return a string listing the found types and count for each
+	 */
+	public String printCardinalityOfAny(final Class<?> type) {
+		final Object[] objs = getInstancesOfAny(type);
+		Map<Class<?>, Integer> card = classCardinality(objs);
+		final StringBuilder b = new StringBuilder(card.size() * 20);		
+		for(Map.Entry<Class<?>, Integer> entry: card.entrySet()) {
+			if(b.length() > 0) {
+				b.append("\n");
+			}
+			b.append(entry.getKey().getName()).append(":").append(entry.getValue());			
+		}
+		return b.toString();		
 	}
 	
-		
+	/** The sys prop for the library path */
+	public static final String SYS_LIB_PATH = "java.library.path";
+	
+	public static final String PATH_SEPARATOR = File.pathSeparator;
+	
+	/**
+	 * Appends the passed path to the java library search path
+	 * @param path The path to append
+	 */
+	public void appendToLibPath(final String path) {
+		final String current = System.getProperty(SYS_LIB_PATH);
+		final String newPath = (current==null ? "" : (current + PATH_SEPARATOR)) + path;
+		System.setProperty(SYS_LIB_PATH, newPath);
+		resetLibPath();
+	}
+	
+	private static void resetLibPath() {
+		try {
+			final Field fieldSysPath = ClassLoader.class.getDeclaredField( "sys_paths" ); 
+			fieldSysPath.setAccessible( true ); 
+			fieldSysPath.set( null, null ); 
+		} catch (Exception ex) {
+			throw new RuntimeException(ex);
+		}
+	}
+	
+
+	//============================================================================
+	//	Native JVMTI calls
+	//============================================================================
 	private static native int countExactInstances0(Class<?> klass);	
 	private static native int countInstances0(Class<?> klass, long tag, int maxInstances);
 	private static native Object[] getExactInstances0(Class<?> klass, long tag, int maxInstances);
